ffdec: zero copy support for edges
Increase the frame size to add padding, and use a crop event to crop out
the edges in the display.  This avoids the need to memcpy for codecs that
don't support CODEC_FLAG_EMU_EDGE.  Currently only enabled for ffvp8, but
could be enabled for other codecs which still disable direct rendering
(VP5, VP6*, H264).